requirements.txt
aiodns==1.1.1
aiohttp==3.4.4
arrow==0.12.1
cchardet==2.1.1
cryptography==2.3.1
requests==2.19.1
furl==1.2
typing==3.6.4
wheel==0.32.2
factory-boy==2.11.1

Django==2.1.3
psycopg2>=2.7.5
psycopg2-binary>=2.7.6.1
django-extensions==2.1.3
djangorestframework==3.8.2
djangorestframework-recursive>=0.1.2
djangorestframework-jwt>=1.11.0
django-bootstrap4==0.0.6
django-cors-headers==2.4.0
django-mptt>=0.9.1
Pillow==5.2.0

django-rest-auth==0.9.3
django-allauth==0.38.0

geopy==1.16.0

